backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
google-generativeai==0.8.3
sentence-transformers==3.0.1
faiss-cpu==1.12.0
numpy==1.26.4
pydantic==2.8.2
